% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FiMAP-package.R
\docType{package}
\name{FiMAP-package}
\alias{FiMAP-package}
\title{Fast Identity-by-Descent Mapping Test}
\description{
Perform efficient identity-by-descent (IBD) mapping test using IBD segments to identify genomic regions associated with complex traits in large biobank-scale cohorts.
}
\author{
Han Chen \email{Han.Chen.2@uth.tmc.edu}
}
