% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clusters.R
\name{get_clusters}
\alias{get_clusters}
\title{Get clusters from a dsCMatrix}
\usage{
get_clusters(x)
}
\arguments{
\item{x}{The dsCMatrix}
}
\value{
An index vector showing which observations are in the same cluster
}
\description{
Get clusters from a dsCMatrix
}
