% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_local_ibd.R
\name{get_local_ibd}
\alias{get_local_ibd}
\title{Calculate local IBD proportions from IBD segment summary files}
\usage{
get_local_ibd(
  ids,
  infiles,
  chr.length,
  window.size,
  id1.col = 2,
  id2.col = 3,
  start.col,
  end.col,
  is.integer = FALSE,
  cut.off = 0,
  ncores = 1
)
}
\arguments{
\item{ids}{The full ID vector}

\item{infiles}{A character vector of input file names}

\item{chr.length}{Total length of the chromosome (can be in cM or bp)}

\item{window.size}{Chunking window size (can be in cM or bp)}

\item{id1.col}{Column number of individual 1 ID in \code{infiles}}

\item{id2.col}{Column number of individual 2 ID in \code{infiles}}

\item{start.col}{Column number of start positions for each IBD segment in \code{infiles}}

\item{end.col}{Column number of end positions for each IBD segment in \code{infiles}}

\item{is.integer}{A logical switch: whether the start and end positions of each IBD segment are integers (default = FALSE for cM; use \code{is.integer = TRUE} for bp)}

\item{cut.off}{Threshold to zero out small IBD values (default = 0)}

\item{ncores}{A positive integer indicating the number of cores to be used in parallel computing (default = 1).}
}
\value{
A list of length equal to the number of windows. Each window is also a list with the following three components:
\item{start}{The start position}
\item{end}{The end position}
\item{matrix}{A dsCMatrix}
}
\description{
Calculate local IBD proportions from IBD segment summary files
}
