% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmkin2randomvec.R
\name{glmmkin2randomvec}
\alias{glmmkin2randomvec}
\title{Create random vectors for a glmmkin object}
\usage{
glmmkin2randomvec(
  obj,
  Z = NULL,
  N.randomvec = 1000,
  group.idx = NULL,
  cluster.idx = NULL,
  robust = FALSE
)
}
\arguments{
\item{obj}{The glmmkin object}

\item{Z}{A list of design matrices for the random effects. The length must match the number of variance components}

\item{N.randomvec}{The number of random vectors to generate}

\item{group.idx}{A length N index vector showing which observation belongs to which variance group, for heteroscedastic linear mixed models (default = NULL for homoscedastic linear mixed models)}

\item{cluster.idx}{A length N index vector showing which observation belongs to which cluster (default = NULL for no clusters)}

\item{robust}{A logical switch: whether robust variance should be used (default = FALSE)}
}
\value{
A list of class glmmkin.randomvec
\item{theta}{Variance estimates, inherited from the glmmkin object}
\item{scaled.residuals}{Scaled residuals, inherited from the glmmkin object}
\item{random.vectors}{An N by N.randomvec matrix of the random vectors generated}
\item{ncovar}{Number of fixed-effects covariates (including the intercept) from the glmmkin object}
\item{id_include}{The ID vector of included samples, inherited from the glmmkin object}
}
\description{
Create random vectors for a glmmkin object
}
