% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibd_mapping_batch.R
\name{ibd_mapping_batch}
\alias{ibd_mapping_batch}
\title{Conduct IBD mapping for multiple local IBD matrices in batches}
\usage{
ibd_mapping_batch(
  glmmkin.randomvec.obj,
  Sigma = NULL,
  offset = NULL,
  ibd.ids,
  ibd.file.list,
  ncores = 1
)
}
\arguments{
\item{glmmkin.randomvec.obj}{The glmmkin.randomvec object}

\item{Sigma}{An N by N sparse covariance matrix for the fitted glmmkin (default = NULL)}

\item{offset}{An N.randomvec by N.randomvec dense matrix computed as the cross-product of Sigma with glmmkin.randomvec matrix of random vectors on both sides (default = NULL)}

\item{ibd.ids}{The ID vector for row and column names of the local IBD matrices}

\item{ibd.file.list}{A character vector of input file names for local IBD matrices}

\item{ncores}{A positive integer indicating the number of cores to be used in parallel computing (default = 1).}
}
\value{
A data.frame
\item{chr}{The chromosome}
\item{start}{The start position}
\item{end}{The end position}
\item{n.nonzero}{Number of non-zero entries in the upper triangle of the local IBD matrix}
\item{p.value}{Finite-sample p-value}
\item{p.value.asymptotic}{Asymptotic p-value}
}
\description{
Conduct IBD mapping for multiple local IBD matrices in batches
}
