% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makepd_add_diag.R
\name{makepd_add_diag}
\alias{makepd_add_diag}
\title{Make a positive definite IBD matrix}
\usage{
makepd_add_diag(ibd, tol = 1e-05)
}
\arguments{
\item{ibd}{The raw IBD matrix}

\item{tol}{Relative tolerance added to the diagonal entries compared to the maximum value of each column (default = 1e-5)}
}
\value{
The positive definite IBD matrix
}
\description{
Make a positive definite IBD matrix
}
