% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_global_ibd.R
\name{process_global_ibd}
\alias{process_global_ibd}
\title{Calculate pairwise global IBD proportions from IBD segment summary files}
\usage{
process_global_ibd(
  ids,
  infiles,
  total.length = 3391.3543292972,
  outfile.prefix,
  id1.col = 2,
  id2.col = 3,
  start.col,
  end.col,
  is.integer = FALSE,
  ncores = 1
)
}
\arguments{
\item{ids}{The full ID vector}

\item{infiles}{A character vector of input file names}

\item{total.length}{Total length of the genome (can be in cM or bp; default = 3391.3543292972 cM)}

\item{outfile.prefix}{Prefix of output file names. The output files are indexed from 1 to \code{length(infiles)} and gzipped.}

\item{id1.col}{Column number of individual 1 ID in \code{infiles}}

\item{id2.col}{Column number of individual 2 ID in \code{infiles}}

\item{start.col}{Column number of start positions for each IBD segment in \code{infiles}}

\item{end.col}{Column number of end positions for each IBD segment in \code{infiles}}

\item{is.integer}{A logical switch: whether the start and end positions of each IBD segment are integers (default = FALSE for cM; use \code{is.integer = TRUE} for bp)}

\item{ncores}{A positive integer indicating the number of cores to be used in parallel computing (default = 1).}
}
\description{
Calculate pairwise global IBD proportions from IBD segment summary files
}
