% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_global_ibd_combine.R
\name{process_global_ibd_combine}
\alias{process_global_ibd_combine}
\title{Assemble pairwise global IBD proportions into a sparse symmetric matrix}
\usage{
process_global_ibd_combine(ids, infile, cut.off = 0)
}
\arguments{
\item{ids}{The full ID vector}

\item{infile}{The input file name}

\item{cut.off}{Threshold to zero out small IBD values (default = 0)}
}
\value{
A dsCMatrix
}
\description{
Assemble pairwise global IBD proportions into a sparse symmetric matrix
}
